feat(1- rota delete implementada; 2- condição aplicada(deleta apenas se tiver status pendente))

+++ src/main/java/com/teste/api/services/PagamentosService.java
@@ -1,7 +1,10 @@
 package com.teste.api.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+
+import com.teste.api.controller.config.exception.ApiExceptionMessage;
 import com.teste.api.dto.PagamentoDTO;
 import com.teste.api.dto.ProcessamentoDTO;
 import com.teste.api.model.PagamentoModel;
@@ -38,15 +41,24 @@ public class PagamentosService {
         repository.save(new PagamentoModel(req));
     }
 
-    public void deletarPagamento(Long id) {
-        // PagamentoModel dadosPagamentos = repository.findById(id);
-        /*
-         * if (id == null || !) {
-         * throw new Error("error");
-         * }
-         */
+    public PagamentoModel deletarPagamento(Long id) throws ApiExceptionMessage {
+        PagamentoModel pagamento = repository.findById(id).orElse(null);
+
+        if (pagamento == null) {
+            throw new ApiExceptionMessage(HttpStatus.BAD_REQUEST, "ID não encontrado");
+        }
+
+        System.out.println(pagamento.getStatusPagamento());
+        if (!"PENDENTE".equals(pagamento.getStatusPagamento().toString())) {
+            throw new ApiExceptionMessage(HttpStatus.BAD_REQUEST,
+                    "Somente pagamentos pendentes podem ser deletados.");
+        }
+
+        System.out.println(pagamento.getStatusPagamento());
 
         repository.deleteById(id);
+
+        return pagamento;
     }
 
     public void atualizarPagamento(Long id, ProcessamentoDTO req) {
